feat(forum): refactor reaction handling in PostCard component
- Replaced ReactionSystem with ReactionButton and ReactionSummary components for improved modularity.
- Added ReactionSummary to display reaction counts above actions.
- Implemented ReactionButton for user interaction with reactions, enhancing user experience.
- Removed the deprecated ReactionSystem component to streamline codebase.
